@maxxu05 pointed out that the wavelet detector uses a moving average
Currently the wavelet detector had a bandpass filter after the squaring operation
but the wavelet detector follows basically exactly the Pan Tompkins detector
which uses a moving average of the max width of the QRS complex which is
150ms in the Pan Tompkins paper so I replaced the bandpass filter with the
moving average. That actually makes the detection much worse but now we
have a perfect replication of the Kalidas and Tamil paper.